Move WebAuthnRpServiceRegistrationRepository
